HSEARCH-4505 Add compatibility with Elasticsearch 8.1
