+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/lowlevel/syntax/search/impl/ElasticsearchSearchSyntax.java
@@ -8,6 +8,7 @@ package org.hibernate.search.backend.elasticsearch.lowlevel.syntax.search.impl;
 
 import java.util.List;
 
+import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonPrimitive;
 
@@ -20,5 +21,7 @@ public interface ElasticsearchSearchSyntax {
 	void requestNestedSort(List<String> nestedPathHierarchy, JsonObject innerObject, JsonObject filterOrNull);
 
 	void requestGeoDistanceSortIgnoreUnmapped(JsonObject innerObject);
+
+	JsonElement encodeLongForAggregation(Long value);
 
 }
